requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.0
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
auth-token==0.0.3
boto3==1.35.79
botocore==1.35.79
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dj-rest-auth==6.0.0
Django==4.2.17
django-allauth==65.1.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
fcm-django==2.2.1
firebase-admin==6.6.0
frozenlist==1.5.0
git-filter-repo==2.47.0
google-api-core==2.24.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
msgpack==1.1.0
multidict==6.1.0
pillow==11.0.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.1
stripe==11.2.0
swapper==1.4.0
twilio==9.3.6
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.16.0
